refactor diffing
