+++ packages/backend/src/core/discovery/utils/diffContracts.ts
@@ -1,0 +1,77 @@
+import { diff } from "deep-diff"
+
+import { ContractParameters } from "../types"
+
+export interface FieldDiff {
+    key?: string
+    before?: string
+    after?: string
+  }
+
+export function diffContracts(
+    before: unknown,
+    after: ContractParameters,
+    ignoreInWatchMode: string[],
+  ): FieldDiff[] {
+    const differences = diff(before, after)
+  
+    if (differences === undefined) {
+      return []
+    }
+  
+    const result: FieldDiff[] = []
+  
+    for (const difference of differences) {
+      switch (difference.kind) {
+        case 'N':
+          result.push({
+            key: difference.path?.join('.'),
+            after: JSON.stringify(difference.rhs),
+          })
+          break
+        case 'D':
+          result.push({
+            key: difference.path?.join('.'),
+            before: JSON.stringify(difference.lhs),
+          })
+          break
+        case 'E':
+          result.push({
+            key: difference.path?.join('.'),
+            before: JSON.stringify(difference.lhs),
+            after: JSON.stringify(difference.rhs),
+          })
+          break
+        case 'A':
+          {
+            const r: FieldDiff = {
+              key: `${difference.path?.join('.') ?? 'undefined'}[${
+                difference.index
+              }]`,
+            }
+            if (difference.item.kind === 'N') {
+              r.after = JSON.stringify(difference.item.rhs)
+            }
+            if (difference.item.kind === 'D') {
+              r.before = JSON.stringify(difference.item.lhs)
+            }
+            if (difference.item.kind === 'E') {
+              r.before = JSON.stringify(difference.item.lhs)
+              r.after = JSON.stringify(difference.item.rhs)
+            }
+            result.push(r)
+          }
+          break
+      }
+    }
+  
+    for (const ignored of ignoreInWatchMode) {
+      const index = result.findIndex((r) => r.key?.includes(`values.${ignored}`))
+      if (index > -1) {
+        result.splice(index, 1)
+      }
+    }
+  
+    return result
+  }
+
