+++ packages/backend/src/core/discovery/utils/diffDiscovery.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'earljs'
 
 import { DiscoveryContract } from '../DiscoveryConfig'
 import { ContractParameters } from '../types'
-import { diffDiscovery, diffObjects } from './diffDiscovery'
+import { diffDiscovery } from './diffDiscovery'
 
 describe(diffDiscovery.name, () => {
   const ADDRESS_A = EthereumAddress.random()
@@ -13,7 +13,7 @@ describe(diffDiscovery.name, () => {
 
   const ADMIN = EthereumAddress.random()
   const IMPLEMENTATION = EthereumAddress.random()
-  it('works', () => {
+  it('finds changes, deleted and created contracts', () => {
     const committed: unknown[] = [
       //finds changes
       {
@@ -113,68 +113,6 @@ describe(diffDiscovery.name, () => {
         address: ADDRESS_C,
         type: 'created',
       },
-    ])
-  })
-})
-
-describe(diffObjects.name, () => {
-  const OLD_ADDRESS = EthereumAddress.random()
-  const NEW_ADDRESS = EthereumAddress.random()
-
-  const OLD_ADMIN = EthereumAddress.random()
-  const NEW_ADMIN = EthereumAddress.random()
-
-  const IMPLEMENTATION = EthereumAddress.random()
-
-  it('returns keys of changed fields', () => {
-    const committed = {
-      name: 'A',
-      address: OLD_ADDRESS.toString(),
-      upgradeability: {
-        type: 'EIP1967 proxy',
-        admin: OLD_ADMIN.toString(),
-        implementation: IMPLEMENTATION,
-      },
-      values: {
-        A: true,
-        B: true,
-        C: 1,
-        D: [1, 2, 3],
-        E: 'ignoreMe',
-      },
-    }
-    const discovered: ContractParameters = {
-      name: 'A',
-      address: NEW_ADDRESS,
-      upgradeability: {
-        type: 'EIP1967 proxy',
-        admin: NEW_ADMIN,
-        implementation: IMPLEMENTATION,
-      },
-      values: {
-        A: false,
-        C: 1,
-        D: [1, 2, 3, 4],
-        E: 'ignoreMePlease',
-      },
-    }
-    const ignoreInWatchMode = ['E']
-    const result = diffObjects(committed, discovered, ignoreInWatchMode)
-
-    expect(result).toEqual([
-      {
-        key: 'address',
-        before: JSON.stringify(OLD_ADDRESS),
-        after: JSON.stringify(NEW_ADDRESS),
-      },
-      {
-        key: 'upgradeability.admin',
-        before: JSON.stringify(OLD_ADMIN),
-        after: JSON.stringify(NEW_ADMIN),
-      },
-      { key: 'values.A', before: 'true', after: 'false' },
-      { key: 'values.B', before: 'true' },
-      { key: 'values.D[3]', after: '4' },
     ])
   })
 })
